Added full functionality for removing auctions and creating auctions.
Added a line in the calendar model class that resets the lastAuctionDate
back to default if an auction is cancelled. The argument for allowing
this is that if they had an auction already scheduled then there
previous auctions was more than one year out. by setting it backto
default we remove the issue with the system thinking that the last
auction was actually the one that was cancelled last which could cause
issues.